Enable ReleasePlugin to release itself
I feel kinda dirty by doing this, but on the other hand it is cool that
the release plugin can release itself without depending on a previously
released artifact.
As the comment in project/build.sbt mentions, we avoid the bootstrapping
problem by utilizing this hack. In addition, new releases of the plugin
will always be created by the most recent code rather than a possibly
buggy previous release.